requirements.txt
attrs==19.3.0
Babel==2.7.0
blinker==1.4
branca==0.3.1
certifi==2019.9.11
chardet==3.0.4
Click==7.0
click-plugins==1.1.1
cligj==0.5.0
dominate==2.4.0
Fiona==1.8.9
Flask==1.1.1
Flask-BabelEx==0.9.3
Flask-Bootstrap4==4.0.2
Flask-Mail==0.9.1
Flask-Principal==0.4.0
Flask-Session==0.3.1
Flask-SQLAlchemy==2.4.1
Flask-WTF==0.14.2
folium==0.10.0
GDAL==3.0.1
geopandas==0.6.1
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
munch==2.5.0
numpy==1.17.3
pandas==0.25.2
passlib==1.7.1
pyproj==2.4.0
python-dateutil==2.8.0
pytz==2019.3
requests==2.22.0
Rtree==0.8.3
Shapely==1.6.4.post2
six==1.12.0
speaklater==1.3
SQLAlchemy==1.3.10
urllib3==1.25.6
visitor==0.1.3
Werkzeug==0.16.0
WTForms==2.2.1
